Public key URL not resolved correctly
The previous version introduced a change that removed the need for a discovery call to find the keys endpoint.
However the concatenation of `issuerUrl + "/v1/keys"` was invalid. This path was mock/spy to avoid making a live call to the discovery endpoint.

This change removes the mock/spy (no longer needed), and adds a test to verify the keys URL is set to the expected `issueUrl + "/v1/keys"`

Fixes: #26